# Improve the performance of redirection