--- a/src/exec.cpp
+++ b/src/exec.cpp
@@ -12,24 +12,90 @@ enum READ_IO_TYPE : int {
     READ_STDERR,
 };
 
-unsigned ReadIO(subprocess_s &process,
-                int read_io_type,
-                char *buf, const unsigned buf_size,
-                noex::string& str, const size_t str_size) noexcept {
-    unsigned read_size = 0;
-    if (read_io_type == READ_STDOUT) {
-        read_size = subprocess_read_stdout(&process, buf, buf_size);
-    } else {
-        read_size = subprocess_read_stderr(&process, buf, buf_size);
+#define BUF_SIZE 65536
+#define LAST_CHARS_MAX_LEN 2048
+
+class RedirectContext {
+ private:
+#ifdef _WIN32
+    HANDLE m_file;
+#else
+    FILE* m_file;
+#endif
+    int m_io_type;
+    char m_buf[BUF_SIZE + 1];
+    noex::string m_last_chars;  // Stores the last characters
+    bool m_use_utf8_on_windows;
+
+    noex::string TruncateStr(const noex::string& str, size_t size) noexcept {
+        if (str.size() > size)
+            return "..." + str.substr(str.size() - size, size);
+        return str;
     }
 
-    buf[read_size] = 0;
-    str += buf;
-    if (str.length() > str_size * 2)
-        str.erase(0, str.length() - str_size);
+ public:
+    explicit RedirectContext(int read_io_type, int use_utf8_on_windows) noexcept :
+            m_use_utf8_on_windows(use_utf8_on_windows),
+            m_io_type(read_io_type), m_last_chars() {
+    #ifdef _WIN32
+        if (read_io_type == READ_STDOUT)
+            m_file = GetStdHandle(STD_OUTPUT_HANDLE);
+        else
+            m_file = GetStdHandle(STD_ERROR_HANDLE);
+    #else
+        if (read_io_type == READ_STDOUT)
+            m_file = stdout;
+        else
+            m_file = stderr;
+    #endif
+    }
 
-    return read_size;
-}
+    void RedirectOutput(subprocess_s &process) noexcept {
+        unsigned read_size = 0;
+        while (1) {
+            // Read outputs
+            if (m_io_type == READ_STDOUT)
+                read_size = subprocess_read_stdout(&process, m_buf, BUF_SIZE);
+            else
+                read_size = subprocess_read_stderr(&process, m_buf, BUF_SIZE);
+            m_buf[read_size] = 0;
+
+            if (!read_size)
+                break;
+
+            // Store last characters
+            m_last_chars += m_buf;
+            if (m_last_chars.length() > LAST_CHARS_MAX_LEN * 2)
+                m_last_chars.erase(0, m_last_chars.length() - LAST_CHARS_MAX_LEN);
+
+            // Redirect to console
+        #ifdef _WIN32
+            DWORD written;
+            if (m_use_utf8_on_windows) {
+                noex::wstring wout = UTF8toUTF16(m_buf);
+                WriteConsoleW(m_file, wout.c_str(), static_cast<DWORD>(wout.size()), &written, NULL);
+            } else {
+                WriteFile(m_file, m_buf, static_cast<DWORD>(read_size), &written, NULL);
+            }
+        #else  // _WIN32
+        #ifdef __TUW_UNIX__
+            Log(m_buf);
+        #endif
+            fwrite(m_buf, sizeof(char), read_size, m_file);
+        #endif  // _WIN32
+        }
+    }
+
+    noex::string GetLastChars() noexcept {
+        return TruncateStr(m_last_chars, LAST_CHARS_MAX_LEN);
+    }
+
+    noex::string GetLine() noexcept {
+        noex::string last_line = GetLastLine(m_last_chars);
+        last_line = TruncateStr(last_line, LAST_CHARS_MAX_LEN);
+        return last_line;
+    }
+};
 
 void DestroyProcess(subprocess_s &process,
                     int *return_code, noex::string &err_msg) noexcept {
@@ -37,39 +103,6 @@ void DestroyProcess(subprocess_s &process,
         *return_code = -1;
         err_msg = "Failed to manage subprocess.\n";
     }
-}
-
-void RedirectOutput(FILE* out, const char* buf,
-    unsigned read_size,
-    bool use_utf8_on_windows) noexcept {
-    if (!read_size)
-        return;
-
-    // Note: We shouldn't use printf as it's too slow for redirection.
-#ifdef _WIN32
-    HANDLE file;
-    DWORD written;
-    if (out == stdout)
-        file = GetStdHandle(STD_OUTPUT_HANDLE);
-    else
-        file = GetStdHandle(STD_ERROR_HANDLE);
-    if (use_utf8_on_windows) {
-        noex::wstring wout = UTF8toUTF16(buf);
-        WriteConsoleW(file, wout.c_str(), wout.size(), &written, NULL);
-    } else {
-        // ANSI code page (It might not be UTF8)
-        WriteFile(file, buf, read_size, &written, NULL);
-    }
-#else
-    Fwrite(out, buf, read_size);
-#endif
-}
-
-
-inline noex::string TruncateStr(const noex::string& str, size_t size) noexcept {
-    if (str.size() > size)
-        return "..." + str.substr(str.size() - size, size);
-    return str;
 }
 
 ExecuteResult Execute(const noex::string& cmd,
@@ -116,46 +149,29 @@ ExecuteResult Execute(const noex::string& cmd,
     if (0 != result)
         return { -1, "Failed to create a subprocess.\n", ""};
 
-    const unsigned BUF_SIZE = 65536;  // 64KB
-    const size_t LAST_LINE_MAX_LEN = 1024;
-    const size_t ERR_MSG_MAX_LEN = 2048;
-    char out_buf[BUF_SIZE + 1];
-    char err_buf[BUF_SIZE + 1];
-    noex::string last_line;
-    noex::string err_msg;
-    unsigned out_read_size = 0;
-    unsigned err_read_size = 0;
+    RedirectContext stdout_context(READ_STDOUT, use_utf8_on_windows);
+    RedirectContext stderr_context(READ_STDERR, use_utf8_on_windows);
 
     do {
-        do {
-            out_read_size = ReadIO(process, READ_STDOUT,
-                                out_buf, BUF_SIZE, last_line, LAST_LINE_MAX_LEN);
-            RedirectOutput(stdout, out_buf, out_read_size, use_utf8_on_windows);
-        } while (out_read_size);
-        do {
-            err_read_size = ReadIO(process, READ_STDERR,
-                            err_buf, BUF_SIZE, err_msg, ERR_MSG_MAX_LEN);
-            RedirectOutput(stderr, err_buf, err_read_size, use_utf8_on_windows);
-        } while (err_read_size);
+        stdout_context.RedirectOutput(process);
+        stderr_context.RedirectOutput(process);
 #ifdef _WIN32
         Sleep(10);  // wait 10ms
 #else
-        nanosleep(&ten_ms, nullptr);
+        nanosleep(&ten_ms, nullptr);  // wait 10ms
 #endif
     } while (subprocess_alive(&process));
 
     // Sometimes stdout and stderr still have unread characters
-    out_read_size = ReadIO(process, READ_STDOUT, out_buf, BUF_SIZE, last_line, LAST_LINE_MAX_LEN);
-    err_read_size = ReadIO(process, READ_STDERR, err_buf, BUF_SIZE, err_msg, ERR_MSG_MAX_LEN);
-    RedirectOutput(stdout, out_buf, out_read_size, use_utf8_on_windows);
-    RedirectOutput(stderr, err_buf, err_read_size, use_utf8_on_windows);
+    stdout_context.RedirectOutput(process);
+    stderr_context.RedirectOutput(process);
+
+    // Get buffered characters from stdout and stderr
+    noex::string last_line = stdout_context.GetLine();
+    noex::string err_msg = stderr_context.GetLastChars();
 
     int return_code;
     DestroyProcess(process, &return_code, err_msg);
-
-    last_line = GetLastLine(last_line);
-    last_line = TruncateStr(last_line, LAST_LINE_MAX_LEN);
-    err_msg = TruncateStr(err_msg, ERR_MSG_MAX_LEN);
 
 #ifdef _WIN32
     if (!use_utf8_on_windows)